feat(seer): Iterate on the instructions at the top of seer settings pages
