+++ static/gsApp/views/seerAutomation/components/settingsPageTabs.tsx
@@ -16,11 +16,8 @@ export default function SettingsPageTabs() {
 
   const tabs: Array<[string, Path]> = [
     [t('Settings'), `/settings/${organization.slug}/seer/`],
-    [
-      t('Projects (Root Cause Analysis & Agents)'),
-      `/settings/${organization.slug}/seer/projects/`,
-    ],
-    [t('Repos (Code Review)'), `/settings/${organization.slug}/seer/repos/`],
+    [t('Issue Autofix'), `/settings/${organization.slug}/seer/projects/`],
+    [t('Code Review'), `/settings/${organization.slug}/seer/repos/`],
   ];
 
   return (
